broke out big fn into little fn part 1
