+++ lib/genOutputFnMapper.js
@@ -1,115 +1,6 @@
 const sourceFn = require('./sourceFn')
-const BigNumber = require('web3').BigNumber
 const Web3 = require('web3')
 const web3 = new Web3()
-
-let graphQLTypeMapper = {
-  'uint256[]': 'uint256Arr'
-}
-
-const convertTypeToGraphQLName = ({type, name}, index) => {
-  if (name !== '') {
-    return name
-  }
-  let queryName = type
-  if (type.includes('[]')) {
-    queryName = `${type.substring(0, type.length - 2)}Arr_${index}`
-  } else {
-    queryName = `${type}_${index}`
-  }
-  // console.log(queryName)
-  return queryName
-}
-
-const abiOutputToGraphQLType = (input, index) => {
-  const { type } = input
-  // console.log(`output.name = ${input.name}, output.type = ${input.type}`)
-  // const name = `${input.type}_${index}`
-  const name = convertTypeToGraphQLName(input, index)
-  // console.log(`name = ${name}`)
-  if (type.includes('int')) {
-    // let value = (input.includes('[]')) ? 'Value[]' : 'Value'
-    // return { key: 'value', value: value }
-    // let nameTest = graphQLTypeMapper[input.type]
-
-    if (type.includes('[]')) {
-      // console.log('================')
-      // console.log(input)
-      // console.log('================')
-      return { key: 'value', value: '[Value]', name, index }
-    }
-
-    return { key: 'value', value: 'Value', name, index }
-  }
-
-  if (type === 'bool') {
-    // console.log(input)
-    let boolean = (type.includes('[]')) ? 'Boolean[]' : 'Boolean'
-    return { key: 'boolean', value: boolean, name, index }
-  }
-
-  if (type === 'string') {
-    return { key: 'string', value: 'String', name, index }
-  }
-
-  if (type.includes('bytes')) {
-    if (type.includes('[]')) {
-      return { key: 'bytes32', value: '[Bytes]', name, index }
-    }
-    return { key: 'bytes32', value: 'Bytes', name, index }
-  }
-
-  if (type.includes('address')) {
-    if (type.includes('[]')) {
-      return { key: 'address', value: '[String]', name, index }
-    }
-    return { key: 'address', value: 'String', name, index }
-  }
-
-  return new Error(`Unkown abi output type: ${type}`)
-}
-
-// const getOutputType = (input, index) => {
-//   // const { type } = input
-
-//   // console.log('======= 1 ')
-//   // console.log(JSON.stringify(input, null, 2))
-//   // console.log(index)
-//   // console.log('======= 2')
-//   // input.name = `${input.type}_index`
-//   return abiOutputToGraphQLType(input, index)
-// }
-
-const genType = ({ typeName, typeBody }) => {
-  return `
-  type ${typeName} {
-    ${typeBody}
-  }`
-}
-const typesToSchemaLine = (data) => {
-  let tmp = data.types
-              .map(type => `${type.name}: ${type.value}`)
-              .reduce((out, cur, index) => {
-                return (index === 0) ? `${out} ${cur}` : `${out} \n     ${cur}`
-              }, '')
-
-  // if (data.name === 'returnsArrayInt') {
-  //   console.log(JSON.stringify(data, null, 2))
-  //   console.log(tmp)
-  // }
-  return tmp
-}
-
-const topString =
-`
-  type Value {
-    string: String
-    int: Int
-  }
-  type Bytes {
-    raw: String
-    decoded: String
-  }`
 
 const genOutputFuncitonMapper = (input) => {
   return (data) => {
@@ -144,10 +35,6 @@ const genOutputFuncitonMapper = (input) => {
       return tmp1
     } else if (resultsIsArr && inputIsArr) {
       if (input.length === data.length) { // multiple input and out
-        // console.log('inptus match outputs')
-        // console.log(data)
-        // console.log(input)
-
         let tmp1 = input.reduce((out, current, index) => {
           // console.log(current)
           // console.log(index)
@@ -222,19 +109,10 @@ const genOutputFuncitonMapper = (input) => {
   }
 }
 
-const genQueryTypes = ({ artifact, contract }) => {
+const genOutputFnMapper = ({ queryConverter, contract, artifact }) => {
   const { abi } = artifact
-  const cleanData = abi
-                .filter(item => typeof item.outputs !== 'undefined')
-                .map(item => { return { name: item.name, outputs: item.outputs } })
-                .map(item => {
-                  // console.log(item)
-                  item.types = item.outputs.map(abiOutputToGraphQLType)
-                  return item
-                })
-  // console.log(JSON.stringify(cleanData, null, 2))
-  // console.log('----- last')
-  const outputMappers = cleanData
+
+  const outputMappers = queryConverter
                 .map(item => {
                   return { fnName: item.name, outputMapper: genOutputFuncitonMapper(item.types) }
                 })
@@ -244,19 +122,7 @@ const genQueryTypes = ({ artifact, contract }) => {
                 }, {})
   // console.log(outputMappers)
   // console.log('-----')
-  const queryTypes = cleanData
-                .map(item => {
-                  const typeName = `${item.name}` // function name = type name
-                  // console.log(`typeName = ${}`)
-                  const typeBody = typesToSchemaLine(item)
-                  return genType({ typeName, typeBody })
-                })
-                .reduce((output, current) => {
-                  return output + current
-                }, topString)
 
-  // console.log(queryTypes)
-  // console.log('-----')
   const allFns = abi
                   .filter(item => item.type === 'function')
                   .map(item => { return { contract, fnName: item.name } })
@@ -275,13 +141,11 @@ const genQueryTypes = ({ artifact, contract }) => {
                           }, {})
 
   return {
-    queryTypes,
     outputMappers,
     allResolvers
   }
 }
 
 module.exports = {
-  genQueryTypes
+  genOutputFnMapper
 }
-// console.log(outputMappers)
