+++ lib/index.js
@@ -1,15 +1,18 @@
 const { buildSchema } = require('graphql')
-const { genQueryTypes } = require('./createQLType')
 const { genFnLines } = require('./createFnQueryLines')
+const { genQueryConverter } = require('./genQueryConverter')
+const { genQueryTypeSchema } = require('./genQueryTypeSchema')
+const { genOutputFnMapper } = require('./genOutputFnMapper')
 
 const genGraphQlProperties = ({ artifact, contract }) => {
-  const { queryTypes, allResolvers } = genQueryTypes({ artifact, contract })
+  const queryConverter = genQueryConverter({ artifact })
+  const queryTypes = genQueryTypeSchema({ queryConverter })
+  const { allResolvers } = genOutputFnMapper({ queryConverter, artifact, contract })
   const createFnQueryLines = genFnLines({ artifact })
   const baseScheme = `
     ${queryTypes}
     ${createFnQueryLines}
   `
-  // console.log(baseScheme)
   return {
     schema: buildSchema(baseScheme),
     rootValue: allResolvers
